pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nyc-housing-health"
version = "0.1.0"
description = "NYC Child Health Housing Index - Mapping housing conditions and child health outcomes"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "NYC Housing Health Project"}
]
keywords = ["nyc", "housing", "health", "data-analysis", "gis"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: GIS",
]

dependencies = [
    "pandas>=2.0",
    "geopandas>=0.14",
    "numpy>=1.24",
    "requests>=2.31",
    "folium>=0.15",
    "matplotlib>=3.8",
    "seaborn>=0.13",
    "scipy>=1.11",
    "shapely>=2.0",
    "pyproj>=3.6",
    "python-dotenv>=1.0",
    "tqdm>=4.66",
    "sodapy",
]

[project.optional-dependencies]
dev = [
    "jupyter>=1.0",
    "pytest>=7.0",
    "black>=23.0",
    "ruff>=0.1",
]

[project.urls]
"Homepage" = "https://github.com/yourusername/nyc-housing-health"
"Bug Tracker" = "https://github.com/yourusername/nyc-housing-health/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

